Add MPC Quad controller

+++ src/controllers/ddp_casadi_mpc_controller.cpp
@@ -5,13 +5,14 @@ DDPCasadiMPCController::DDPCasadiMPCController(
     std::chrono::duration<double> controller_duration)
     : ddp_config_(ddp_config), kt_(1), look_ahead_index_shift_(1) {
   // parameters from ddp config
-  int N = ddp_config.n();
+  unsigned int N = ddp_config.n();
   double h = ddp_config.h();
   CHECK(h > 0) << "The time step should be greater than 0";
   control_timer_shift_ =
-      std::min(int(std::ceil(controller_duration.count() / h)), N - 1);
+      std::min(uint(std::ceil(controller_duration.count() / h)), N - 1);
   VLOG(1) << "Control timer shift: " << control_timer_shift_;
-  max_look_ahead_index_shift_ = std::ceil(ddp_config.look_ahead_time() / h);
+  max_look_ahead_index_shift_ =
+      uint(std::ceil(ddp_config.look_ahead_time() / h));
   VLOG(1) << "Look ahead index shift: " << max_look_ahead_index_shift_;
   CHECK(max_look_ahead_index_shift_ < N)
       << "Look ahead time should be less than trajectory end time";
@@ -20,7 +21,7 @@ DDPCasadiMPCController::DDPCasadiMPCController(
   xds_.resize(N + 1);
   uds_.resize(N);
   // Times
-  for (int k = 0; k <= N; ++k) {
+  for (unsigned int k = 0; k <= N; ++k) {
     ts_.push_back(k * h);
   }
   max_iters_ = ddp_config.max_iters();
@@ -39,9 +40,9 @@ void DDPCasadiMPCController::resetControls() {
   VLOG(1) << "Resetting Controls";
   // initial controls
   Eigen::VectorXd ui = stationaryControl();
-  int N = ddp_config_.n();
+  unsigned int N = ddp_config_.n();
   us_.resize(N);
-  for (int i = 0; i < N; ++i) {
+  for (unsigned int i = 0; i < N; ++i) {
     us_.at(i) = ui;
   }
   resetDDP();
@@ -56,14 +57,14 @@ void DDPCasadiMPCController::setMaxIters(int iters) {
 
 int DDPCasadiMPCController::getMaxIters() const { return max_iters_; }
 
-void DDPCasadiMPCController::rotateControls(int shift_length) {
-  int N = us_.size();
+void DDPCasadiMPCController::rotateControls(unsigned int shift_length) {
+  unsigned long N = us_.size();
   // Rotate controls by control timer shift
   // in proto file (Default 50 Hz)
-  for (int i = 0; i < N - shift_length; ++i) {
+  for (unsigned int i = 0; i < N - shift_length; ++i) {
     us_[i] = us_[i + shift_length];
   }
-  for (int i = N - shift_length; i < N; ++i) {
+  for (unsigned long i = N - shift_length; i < N; ++i) {
     us_[i] = us_[N - 1];
   }
 }
@@ -74,11 +75,11 @@ bool DDPCasadiMPCController::runImplementation(MPCInputs<StateType> sensor_data,
   bool result = true;
   loop_timer_.loop_start();
   boost::mutex::scoped_lock lock(copy_mutex_);
-  int N = ddp_config_.n();
+  unsigned int N = ddp_config_.n();
   double h = ddp_config_.h();
   double t0 = sensor_data.time_since_goal;
   // Get MPC Reference from high level reference trajectory
-  for (int i = 0; i <= N; ++i) {
+  for (unsigned int i = 0; i <= N; ++i) {
     double t = t0 + i * h;
     std::pair<StateType, ControlType> state_control_pair = goal->atTime(t);
     xds_.at(i) = state_control_pair.first;
@@ -95,7 +96,7 @@ bool DDPCasadiMPCController::runImplementation(MPCInputs<StateType> sensor_data,
   ddp_->Update();
   double J = 1e6; // Assume start cost is some large value
   // Run MPC Iterations
-  for (int i = 0; i < max_iters_; ++i) {
+  for (unsigned int i = 0; i < max_iters_; ++i) {
     ddp_->Iterate();
     // Check for convergence
     if (std::abs(ddp_->J - J) < ddp_config_.min_cost_decrease()) {
